Implemented .env file to allow local testing e.g. SEO, etc
Hard coding the production url in the config > url in the docusaurus.config file results in paths being processed with the production url e.g. SEO image is set with the production url, making it difficult to test in a local environment. Using .env.local file ensures the path generated by docusaurus as part of its build process when testing in the dev environment uses the DEPLOY_URL stated in the .env.local file.